+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- `updateItem` now receives an `Partial<Item>` object as argument.
+
 ## [0.2.1] - 2019-09-05
 
 ### Changed